Modifica la implementación del objeto de configuración para hacerlo más sencillo

+++ src/config/config.js
@@ -11,19 +11,29 @@ let singleton = null;
  * Manipula la configuración del programa almacenada en el Drive del usuaario.
  * Se utiliza el patrón Singleton para devolver siempre el mismo objeto
  *
+ * @param {Object} auth: Objeto de autenticación al que se asocia la configuración.
  * @param {String} name: Nombre del fichero de configuración.
  */
-function Config(name) {
+function Config(auth, name) {
    if(singleton) return singleton;
    singleton = this;
 
    Object.defineProperties(this, {
+      "name": {
+         enumerable: true,
+         writable: false,
+         value: name
+      },
+      "auth": {
+         enumerable: true,
+         writable: false,
+         value: auth
+      },
       "_id": {
          enumerable: false,
          writable: true,
          value: null
       },
-      "name": { value: name || "config.json" },
       "_content": {
          enumerable: false,
          writable: true,
@@ -32,33 +42,24 @@ function Config(name) {
    });
 }
 
-
 Object.defineProperties(Config.prototype, {
    /**
     * Identificador del fichero de configuración en el Drive.
     */
-   id: { 
+   id: {
+      get() { return this._id; }
+   },
+   content: {
+      get() { return this._content; }
+   },
+   status: {
       get() {
-         if(!this._id) this._id = getFileID(this.name);
-         return this._id;
+         if(this.content === null) return "UNAVAILABLE";
+         else if(Object.keys(this.content).length === 0) return "NOCONFIG";
+         else return "READY";
       }
    },
-   /**
-    * Promesa qie indica si el fichero de configuración está vacío.
-    */
-   vacia: {
-      get() { 
-         return new Promise(async resolve => {
-                  const config = await this.get();
-                  resolve(Object.keys(config).length === 0);
-                });
-      }
-   },
-   utils: { value: utils },
-   inicializar: { value: inicializar},
-   seed: {
-      get() { return JSON.parse(default_config); }
-   }
+   utils: { value: utils }
 });
 
 
@@ -68,73 +69,56 @@ Object.defineProperties(Config.prototype, {
  *
  * @returns {Promise}: Promesa con el identificador.
  */
-function getFileID(name) {
-   return new Promise((resolve, reject) => {
-      gapi.client.request({
-         path: "https://www.googleapis.com/drive/v3/files",
-         method: "GET",
-         params: {
-            spaces: "appDataFolder",
-            q: "name = '" + name + "'"
-         }
-      }).then(response => {
-         const files = response.result.files;
-         switch(files.length) {
-            case 0:  // El fichero no existe: se crea
-               console.warn("No hay configuración previa");
-               const params = {
-                        name: name,
-                        parents: ['appDataFolder'],
-                        description: "Configuración de Braulio",
-                        mimeType: "application/json"
-                     };
-               gapi.client.request({
-                  path: "https://www.googleapis.com/drive/v3/files",
-                  method: "POST",
-                  body: params
-               }).then(async response => {
-                     const id = response.result.id;
-                     gapi.client.request({
-                        path: "https://www.googleapis.com/upload/drive/v3/files/" + id,
-                        method: "PATCH",
-                        body: {}
-                     }).then(response => resolve(id));
-                  });
-               break;
-            case 1:  // El fichero existe, se devuelve su ID.
-               resolve(files[0].id);
-               break;
-            default:
-               reject("Hay más de un fichero de configuración");
-         }
-      });
-   });
-}
+Config.prototype.init = function() {
+   if(this.id) return;
 
+   gapi.client.request({
+      path: "https://www.googleapis.com/drive/v3/files",
+      method: "GET",
+      params: {
+         spaces: "appDataFolder",
+         q: "name = '" + this.name + "'"
+      }
+   }).then(response => {
+      const files = response.result.files;
 
-/**
- * Obtiene la configuración.
- *
- * @returns {Promise} Promesa con el contenido.
- */
-Config.prototype.get = function() {
-   return new Promise(async resolve => {
-            if(this._content) resolve(this._content);
-
-            const id = await this.id;
-
+      switch(files.length) {
+         case 0:
+            console.warn("No hay configuración previa");
+            const params = {
+                     name: this.name,
+                     parents: ['appDataFolder'],
+                     description: "Configuración de Braulio",
+                     mimeType: "application/json"
+                  };
             gapi.client.request({
-               path: "https://www.googleapis.com/drive/v3/files/" + id,
+               path: "https://www.googleapis.com/drive/v3/files",
+               method: "POST",
+               body: params
+            }).then(response => {
+                  this._id = response.result.id;
+                  this.set(null);  // Almacenamos una configuración vacía.
+               });
+            break;
+         case 1:
+            // El fichero existe, se devuelve su ID.
+            this._id = files[0].id;
+            gapi.client.request({
+               path: "https://www.googleapis.com/drive/v3/files/" + this.id,
                method: "GET",
                params: {alt: "media"}  // Para que devuelva el fichero y no los metadatos.
             }).then(response => {
                   // Añadimos emails, descripciones, etc.
                   getInfo(response.result).then(response => {
                      this._content = response;
-                     resolve(response);
+                     this.auth.fire("onready", {action: "get"});
                   });
                });
-          });
+            break;
+         default:
+            throw new Error("Hay más de un fichero de configuración");
+      }
+   });
 }
 
 
@@ -142,20 +126,34 @@ Config.prototype.get = function() {
  * Fija la configuración en el Drive.
  *
  * @param {Object} content: Objeto que será el nuevo contenido.
- *
- * @returns {Promise} Promesa con la respuesta a la acción.
  */
 Config.prototype.set = function(content) {
-   return new Promise(async resolve => {
-            gapi.client.request({
-               path: "https://www.googleapis.com/upload/drive/v3/files/" + await this.id,
-               method: "PATCH",
-               body: JSON.stringify(mrproper(JSON.parse(JSON.stringify(content)))),
-            }).then(response => {
-                  this._content = content;
-                  resolve(response);
-               });
-          });
+   if(this.id === null) throw new Error("Configuración no inicializada");
+   const body = content?JSON.stringify(mrproper(JSON.parse(JSON.stringify(content)))):{},
+         prestatus = this.status;
+
+   this._content = content || null;
+
+   gapi.client.request({
+      path: "https://www.googleapis.com/upload/drive/v3/files/" + this.id,
+      method: "PATCH",
+      body: body,
+   }).then(response => {
+         if(content) {
+            // El evento sólo se desencadena si previamente no había configuración.
+            if(prestatus !== "READY") this.auth.fire("onready", {action: "set"});
+            this.auth.fire("savedconfig");
+         }
+         else {
+            const seed = JSON.parse(default_config);
+            // Este método no es enumerable,
+            // así que no aparecerá al recorrer el objeto.
+            Object.defineProperty(seed, "set", {
+               value: config => inicializar.call(this, config || seed)
+            });
+            this.auth.fire("noconfig", {seed: seed});
+         }
+      });
 }
 
 
@@ -165,15 +163,28 @@ Config.prototype.set = function(content) {
  * @returns {Promise} Promesa con la respuesta a la eliminación.
  */
 Config.prototype.remove = function() {
-   return new Promise(async resolve => {
-            gapi.client.request({
-               path: "https://www.googleapis.com/drive/v3/files/" + await this.id,
-               method: "DELETE"
-            }).then(response => {
-                  this._content = this._id = null;
-                  resolve(response);
-               });
-          });
+   if(!this.id) throw new Error("Configuración no inicializada");
+
+   return new Promise(resolve => {
+      gapi.client.request({
+         path: "https://www.googleapis.com/drive/v3/files/" + this.id,
+         method: "DELETE"
+      }).then(response => {
+            this.reset();
+            resolve(response);
+         });
+   });
+}
+
+
+/**
+ * Resetea el objeto para dejarlo en el mismo estado
+ * qye cuando se crea (sin identificador ni contenido).
+ *
+ */
+Config.prototype.reset = function() {
+   this._content = this._id = null;
+   this.auth.fire("onreset");
 }
